ML/pyproject.toml
[tool.poetry]
name = "ml"
version = "0.1.0"
description = ""
authors = ["javascript <ntpjc2vinayak@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.1.4"
faiss-cpu = "^1.9.0.post1"
sentence-transformers = "^3.3.1"
transformers = "^4.47.0"
django-cors-headers = "^4.6.0"
python-dotenv = "^1.0.1"
langchain-google-genai = "^2.0.6"
langchain = "^0.3.9"
djangorestframework = "^3.15.2"
nltk = "^3.9.1"
pdfplumber = "^0.11.4"
pymongo = "^4.10.1"
bcrypt = "^4.2.1"
faker = "^33.1.0"
geopy = "^2.4.1"
gensim = "^4.3.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
